+++ test/twin/property_processor_test.rb
@@ -18,13 +18,19 @@ class PropertyProcessorTest < Minitest::Spec
   end
 
   describe "collection" do
+  	let(:twin) { AlbumTwin.new(Album.new("Live!", Artist.new, [Song.new(1), Song.new(2)])) }
   	it "yields twin, index" do
-	    twin   = AlbumTwin.new(Album.new("Live!", Artist.new, [Song.new(1), Song.new(2)]))
-	    
 	    called = []
 	    Disposable::Twin::PropertyProcessor.new(twin.class.definitions.get(:songs), twin).() { |v, i| called << [v.model, i] }
 
 	    called.inspect.must_equal %{[[#<struct PropertyProcessorTest::Song id=1>, 0], [#<struct PropertyProcessorTest::Song id=2>, 1]]}
+  	end
+
+  	it "yields twin" do
+	    called = []
+	    Disposable::Twin::PropertyProcessor.new(twin.class.definitions.get(:songs), twin).() { |v| called << [v.model] }
+
+	    called.inspect.must_equal %{[[#<struct PropertyProcessorTest::Song id=1>], [#<struct PropertyProcessorTest::Song id=2>]]}
   	end
 
   	it "allows nil collection" do