+++ CHANGES.md
@@ -1,7 +1,7 @@
 # 0.4.1
 
 * `Twin#schema` doesn't `extend` at runtime anymore but uses a decorator for `each`.
-* `PropertyProcessor`
+* `PropertyProcessor` now yields `(twin, index)` if its working on a collection.
 
 # 0.4.1